requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
pandas==2.2.3
openpyxl==3.1.5
greenlet
python-dotenv==1.0.1